Moved sidebar to separate file and removed redux Redux wasn't needed after all for the navigation to work, so it's removed from here
