+++ src/navigation/ReduxNavigation.js
@@ -1,3 +1,4 @@
+/*
 import * as ReactNavigation from 'react-navigation';
 import React from 'react';
 import { AppNavigation } from 'react-native';
@@ -17,3 +18,4 @@ function ReduxNavigation (props) {
 const mapStateToProps = state => ({ nav: state.nav });
 export default connect(mapStateToProps)(ReduxNavigation);
 
+*/
